ml-service/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Machine Learning
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4

# Deep Learning (optional)
torch==2.1.1
transformers==4.35.2

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# OCR
pytesseract==0.3.10

# Text Processing
nltk==3.8.1
spacy==3.7.2

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Redis
redis==5.0.1

# Environment
python-dotenv==1.0.0

# Monitoring
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
